Rename active_Record_setup to delayed_job_active_record
